Prevent data being refreshed when keepOpen is set to true.
When keepOpen set to true when using multiple selection, the dropdown stays open after clicking an item, however the pageNumber would be reset to 1 and a new query would be done resulting in loosing all the loaded results that were populated already.
This prevents the pageNumber being reset and a query triggered after selecting an item when keepOpen=true (ie. closeOnSelect=false).